Consolidate test creds helpers into creds modules
